[Refactor]: 유효성 로직 서비스로 이동
